+++ changelog.md
@@ -5,6 +5,7 @@
 ### Major
 
 - Updated Node.js support to `^22.13.0 || ^24.0.0 || >=26.0.0`.
+- Updated the peer dependency [`eslint`](https://npm.im/eslint) to `^10.0.0`.
 - Use the JSDoc tag `@import` for type imports (requires TypeScript v5.5+).
 - The type `ESLintPlugin` is no longer exported from `eslint-plugin-optimal-modules/eslintPluginOptimalModules.mjs`.
 
@@ -15,6 +16,7 @@
 - Updated GitHub Actions CI config.
 - Updated TypeScript config.
 - Updated VS Code workspace settings.
+- Updated the ESLint config.
 
 ## 3.0.0